requirements.txt
# EdgeX SDK依赖（本地SDK目录）
aiohttp>=3.8.0
cryptography>=41.0.0
starkware-crypto>=0.1.0

# 异步和网络
websockets>=11.0.0
asyncio>=3.4.3

# 数据处理
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# 配置和日志
python-dotenv>=1.0.0
loguru>=0.7.0

# 加密和签名
ecdsa>=0.18.0
